+++ odoo/local-src/sf_sale_invoicing/__manifest__.py
@@ -2,13 +2,15 @@
 # Part of senseFly.
 
 {
-    'name': 'Sale invoice fiscal position',
+    'name': 'Sensefly sale invoicing',
     'version': '1.0',
     'author': 'Telmo Santos',
     'website': 'https://www.sensefly.com',
     'category': 'SenseFly',
     'depends': ['sale', 'account_invoice_fiscal_position_update'],
-    'data': [],
+    'data': [
+        'views/menu.xml'
+    ],
     'demo': [],
     'images': ['static/description/icon.png'],
     'installable': True,